Moved utils to utils. Removed CLIErrors from old code. Removed comments in commands.
